fix: refactored the global search []
