--- packages/@sparrow-common/src/components/global-search/index.ts
@@ -1,6 +1,0 @@
-export { default as SearchBar } from "./SearchBar.svelte";
-export { default as SearchSuggestions } from "./SearchSuggestions.svelte";
-export { default as SuggestionTags } from "./SuggestionTags.svelte";
-export { default as RecentItems } from "./RecentItems.svelte";
-export { default as GlobalSearch } from "./GlobalSearch.svelte";
-export { default as NoResults } from "./NoResults.svelte";
