+++ packages/@sparrow-common/src/components/global-search/sub-components/NoResults.svelte
