+++ packages/@sparrow-common/src/components/global-search/sub-components/SearchBar.svelte
