+++ packages/@sparrow-common/src/components/global-search/sub-components/types/types.ts
